fix: update various components to improve functionality and remove debug logs

+++ src/lib/utils/amountHelper.ts
@@ -135,7 +135,7 @@ export function decimalToInputString(locale: Locale, amount: number): string {
 }
 
 export function inputStringToDecimal(value: string): number {
-  const num = Number.parseInt(value);
+  const num = Number.parseFloat(value);
   if (isNaN(num)) return 0;
   return Math.round(num * 100);
 }
